don't appply quartz64 patches The patches for the quartz64 seems to change a lot of stuff, resulting in regressions at least on the pinebook pro.
So don't apply these patches for now,
until the cause of these regressions have been found.

The regressions I have noticed on my Pinebook Pro are:
Red status LED when kernel is running, used to be green.
USB-C charging says it's working, but charging LED is off
and the battery is actually discharging, even though
userspace says it's charging.

Signed-off-by: Dan Johansen <strit@manjaro.org>